repo_github_manager/requirements.txt
PyGithub==1.58.2
python-dotenv==1.0.0
PyYAML==6.0.1
pytest==7.4.0
pytest-mock==3.11.1
coverage==7.3.2 